Move various memory types to uefi-raw and re-export in uefi
